+++ Changes.md
@@ -18,6 +18,8 @@ Version 3.6.1.0 - TBD
  14. Fix quiz drag and drop colour and not working in preview.
  15. Fix content overflow causing horizontal scrollbar when editing a quiz.
  16. Add floating drawer on screens less than 768px wide.
+ 17. Add Alex Gorbatchev's Syntax Highlighter (alexgorbatchev.com/SyntaxHighlighter), version 3.0.83.
+     Licensed under LGPLv3, www.gnu.org/copyleft/lesser.html.
 
 Version 3.6.0.5
   1. Fix core_text scope issue causing exception.